Make the url a bytestring
The URL should be a bytestring rather than a unicode because if its a unicode then the type of msg in httplib _send_output() becomes unicode which in turn tries to concatinate message_body which is a bytestring to msg.
Here if the message_body contains a non ascii character, the concatination fails.
Hence a URL in bytestring will make sure everything remains a bytestring